+++ db/sqlc/querier.go
@@ -13,7 +13,7 @@ type Querier interface {
 	CreatePlayer(ctx context.Context, arg CreatePlayerParams) (sql.Result, error)
 	GetPlayersByCareerYear(ctx context.Context, arg GetPlayersByCareerYearParams) ([]Player, error)
 	GetPlayersOfMostRuns(ctx context.Context) (Player, error)
-	GetPlayersOfMostRunsByCareerEndYear(ctx context.Context, careerEndYear sql.NullInt32) (Player, error)
+	GetPlayersOfMostRunsByCareerEndYear(ctx context.Context, careerEndYear sql.NullInt64) (Player, error)
 }
 
 var _ Querier = (*Queries)(nil)
